fix: 닉네임정보 불러오기 관련 버그 수정2
