+++ src/pages/mypage/ui/UserCalendar.jsx
@@ -4,7 +4,7 @@ import PrevMonth from '../../../assets/icons/common/prevmonth.svg?react';
 import NextMonth from '../../../assets/icons/common/nextmonth.svg?react';
 import { Color } from '../../../components/container/Color';
 import { PuppleButton } from '../../../components/button/Button';
-import Task from './Task';
+import UserTaskList from './UserTaskList';
 
 const UserCalendar = () => {
     const [date, setDate] = useState(new Date());
@@ -82,7 +82,7 @@ const UserCalendar = () => {
                 </Grid>
                 <AddGoogleCalendarButton>구글 캘린더 연동하기</AddGoogleCalendarButton>
             </CalendarWrapper1>
-            <Task selectedDate={selectedDate} />
+            <UserTaskList selectedDate={selectedDate} />
         </CalendarWrapper>
     );
 };
