+++ src/pages/mypage/ui/UserPost.jsx
@@ -4,7 +4,7 @@ import { PuppleButton } from '../../../components/button/Button';
 import Loading from '../../../components/common/Loading';
 import { postAPI } from '../api/postAPI';
 
-const UserPost = () => {
+const UserPost = ({ nickName }) => {
     const [page, setPage] = useState(1);
     const [isLoading, setIsLoading] = useState(false);
     const [posts, setPosts] = useState([]);
@@ -70,7 +70,7 @@ const UserPost = () => {
 
     return (
         <MyPostWrapper>
-            <ExtraBold>내가 쓴 글</ExtraBold>
+            <ExtraBold>{nickName}님이 쓴 글</ExtraBold>
             <Header>
                 <PostRouteButton $isActive={category === 0} onClick={() => setCategory(0)}>
                     질문하기
@@ -124,7 +124,7 @@ const MyPostWrapper = styled.div`
 const ExtraBold = styled.div`
     font-weight: 800;
     font-size: 1.25em;
-
+    color: #8e59ff;
     @media (max-width: 768px) {
         font-size: 1em;
     }
